Update nix & add SDF cursor
Update nix derivations to build with new dependencies, slightly gnarly
as the bindings-freetype-gl has a submodule C dependency and a borked
Setup.hs file which needed patching.

The source is referenced using the same ref as the submodules. There is
a case for using `nih` as the source of truth for future proofing.

`cabal new-run codex` and marvel at the glory of an SDF cursor, or not,
it's probably not that cool unless you're me.