Nix shell updates.
Included changes to cabal.project file so there is now one for working
in a nix-shell. As the only dependency that I couldn't make work in
the shell.nix was the hsc2hs binary that despite my best efforts would
not be overridden.

Readded some missing dependencies to the fontconfig cabal for testing.

Fixed a bug in the addpkg.sh that double referenced the 'ref', causing
errors.

Added some extra things to the gitignore.