Update nixpkgs pin.
Overhaul the shell.nix to remove overrides and patches we no longer need.

Delete the redundant files from the .nix folder.

Update some of the `.hsc` files because of a new parameter in one of
the `inline-c` functions.

Update the `hsc2hs` commit.

Add configure to the top-level `Makefile`.

Remove the need for `tasty-discover` in `language-server`.